refactor: separate human-specific UI
